document size needs to be long for larger files
git-svn-id: http://owltools.googlecode.com/svn/trunk@2090 18f1da76-1bb4-b526-5913-e828fe20442d

+++ OWLTools-Annotation/src/main/java/owltools/gaf/metadata/AnnotationDocumentMetadata.java
@@ -7,6 +7,6 @@ public class AnnotationDocumentMetadata {
 	public int annotatedEntityCount; // aka Gene Products Annotated
 	public int annotationCount;
 	public int annotationCountExcludingIEA;
-	public int gafDocumentSizeInBytes;
+	public long gafDocumentSizeInBytes;
 
 }
